# [quantization] Factor unnecesary work out of add inner loop